# Song/api/swagger.yaml
basePath: /
definitions:
  models.Song:
    properties:
      id:
        type: string
      title:
        type: string
      artist:
        type: string
    type: object
info:
  contact:
    email: ibrahima.coulibaly@etu.uca.fr
    name: Ibrahima Coulibaly
  description: API to manage songs.
  title: Projet_Middleware/Song
  version: 1.0.0
paths:
  /songs:
    get:
      description: Get songs.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Song'
            type: array
        "500":
          description: Something went wrong
      summary: Get songs.
      tags:
      - songs
    post:
      description: Create a new song.
      parameters:
      - name: song
        in: body
        description: The song object to be created
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "500":
          description: Something went wrong
  /songs/{id}:
    get:
      description: Get a song.
      parameters:
      - description: Song UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Get a song.
      tags:
      - songs
    put:
      description: Update a specific song by ID.
      parameters:
      - description: Song UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      - name: song
        in: body
        description: Updated song data
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Update a specific song by ID.
      tags:
      - songs
    delete:
      description: Delete a specific song by ID.
      parameters:
      - description: Song UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Delete a specific song by ID.
      tags:
      - songs
schemes:
- http
swagger: "2.0"